Update policies.json
Updated Security devices. This is after Ubuntu (Linux Mint) testing. The end-user will need to set their locations. Use the find command for the file.so they need and change the location in their policies.json file. Maybe in the future we can have a RHEL policy on this repo as well.